+++ src/table.c
@@ -31,7 +31,7 @@ ecs_table_column_t* new_columns(
     ecs_type_t type)
 {
     ecs_table_column_t *result = ecs_os_calloc(
-        sizeof(ecs_table_column_t), ecs_vector_count(type) + 1);
+        sizeof(ecs_table_column_t), (uint64_t)ecs_vector_count(type) + 1);
 
     ecs_assert(result != NULL, ECS_OUT_OF_MEMORY, NULL);
 
@@ -350,7 +350,7 @@ int16_t ecs_table_dim(
     return 0;
 }
 
-uint64_t ecs_table_count(
+uint32_t ecs_table_count(
     ecs_table_t *table)
 {
     ecs_assert(table != NULL, ECS_INTERNAL_ERROR, NULL);
